backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_groq
langchain_community
langchain_google_genai
pydantic
python-multipart
pypdf
faiss-cpu
